+++ packages/pwa-kit-dev/bin/pwa-kit-dev.js
@@ -6,19 +6,14 @@
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 const p = require('path')
-const fs = require('fs')
+const fse = require('fs-extra')
 const program = require('commander')
 const isEmail = require('validator/lib/isEmail')
 const {execSync: _execSync} = require('child_process')
 const scriptUtils = require('../scripts/utils')
-const sh = require('shelljs')
 const uploadBundle = require('../scripts/upload.js')
 const pkg = require('../package.json')
 const {getConfig} = require('pwa-kit-runtime/utils/ssr-config')
-
-const pkgRoot = p.join(__dirname, '..')
-
-const projectPkg = require(p.join(process.cwd(), 'package.json'))
 
 const execSync = (cmd, opts) => {
     const defaults = {stdio: 'inherit'}
@@ -26,7 +21,9 @@ const execSync = (cmd, opts) => {
 }
 
 const main = () => {
+    const pkgRoot = p.join(__dirname, '..')
     process.env.CONTEXT = process.cwd()
+
     program.description(
         [
             `PWA Kit Dev`,
@@ -88,10 +85,7 @@ const main = () => {
         .action(({user, key}) => {
             try {
                 const settingsPath = scriptUtils.getSettingsPath()
-                fs.writeFileSync(
-                    settingsPath,
-                    JSON.stringify({username: user, api_key: key}, null, 4)
-                )
+                fse.writeJson(settingsPath, {username: user, api_key: key}, {spaces: 4})
                 console.log(`Saved Managed Runtime credentials to "${settingsPath}".`)
             } catch (e) {
                 console.error('Failed to save credentials.')
@@ -106,42 +100,66 @@ const main = () => {
         .addOption(
             new program.Option('--inspect', 'enable debugging with --inspect on the node process')
         )
-        .action(({inspect}) => {
-            execSync(`node${inspect ? ' --inspect' : ''} ${p.join(process.cwd(), 'app', 'ssr.js')}`)
+        .addOption(new program.Option('--noHMR', 'disable the client-side hot module replacement'))
+        .action(({inspect, noHMR}) => {
+            execSync(
+                `node${inspect ? ' --inspect' : ''} ${p.join(process.cwd(), 'app', 'ssr.js')}`,
+                {
+                    env: {
+                        ...process.env,
+                        ...(noHMR ? {HMR: 'false'} : {})
+                    }
+                }
+            )
         })
 
     program
         .command('build')
+        .addOption(
+            new program.Option(
+                '-b, --buildDirectory <buildDirectory>',
+                'the directory where your project should be built'
+            )
+                .default(p.join(process.cwd(), 'build'), './build')
+                .env('PWA_KIT_BUILD_DIR')
+        )
         .description(`build your app for production`)
-        .action(() => {
+        .action(({buildDirectory}) => {
             const webpack = p.join(require.resolve('webpack'), '..', '..', '..', '.bin', 'webpack')
             const projectWebpack = p.join(process.cwd(), 'webpack.config.js')
-            const webpackConf = fs.existsSync(projectWebpack)
+            const webpackConf = fse.pathExistsSync(projectWebpack)
                 ? projectWebpack
                 : p.join(__dirname, '..', 'configs', 'webpack', 'config.js')
-            const silentState = sh.config.silent
-            sh.rm('-rf', './build')
+            fse.emptyDirSync(buildDirectory)
             execSync(`${webpack} --config ${webpackConf}`, {
                 env: {
                     NODE_ENV: 'production',
-                    ...process.env
+                    ...process.env,
+                    // Command option overrides the env var, so we must continue that pattern
+                    PWA_KIT_BUILD_DIR: buildDirectory
                 }
             })
 
             // Copy the project `package.json` into the build folder.
-            sh.cp(p.resolve('package.json'), './build/package.json')
+            fse.copyFileSync('package.json', p.join(buildDirectory, 'package.json'))
 
             // Copy config files.
             const config = p.resolve('config')
-            if (fs.existsSync(config)) {
-                sh.cp('-R', config, './build')
-                sh.config.silent = true
-                sh.rm('./build/config/local.*')
-                sh.config.silent = silentState
+            if (fse.pathExistsSync(config)) {
+                fse.copySync(
+                    config,
+                    p.join(buildDirectory, 'config'),
+                    (file) => !p.basename(file).startsWith('local.')
+                )
             }
 
-            // This file is required by MRT, for historical reasons.
-            sh.touch('./build/loader.js')
+            const loader = p.join(buildDirectory, 'loader.js')
+            if (!fse.pathExistsSync(loader)) {
+                fse.outputFileSync(
+                    loader,
+                    '// This file is required by Managed Runtime for historical reasons.\n'
+                )
+            }
         })
 
     program
@@ -156,14 +174,19 @@ const main = () => {
         .addOption(
             new program.Option(
                 '-m, --message <message>',
-                "a message to include along with the uploaded bundle in Managed Runtime (default: '<git branch>:<git commit hash>')"
+                'a message to include along with the uploaded bundle in Managed Runtime'
             )
+                // The default message is loaded dynamically as part of `uploadBundle(...)`
+                .default(undefined, '<git branch>:<git commit hash>')
         )
         .addOption(
             new program.Option(
                 '-s, --projectSlug <projectSlug>',
-                "a project slug that differs from the name property in your project's package.json (default: the 'name' key from the package.json)"
-            ).default(projectPkg.name)
+                "a project slug that differs from the name property in your project's package.json"
+            )
+                // We load the slug from the package.json by default, but we don't want to do that
+                // unless we need to, so it is loaded conditionally in the action implementation
+                .default(undefined, "the 'name' key from the package.json")
         )
         .addOption(
             new program.Option(
@@ -178,8 +201,23 @@ const main = () => {
 
             const mobify = getConfig() || {}
 
+            if (!projectSlug) {
+                try {
+                    // Using the full path isn't strictly necessary, but results in clearer errors
+                    const projectPkg = p.join(process.cwd(), 'package.json')
+                    const {name} = fse.readJsonSync(projectPkg)
+                    if (!name) throw new Error(`Missing "name" field in ${projectPkg}`)
+                    projectSlug = name
+                } catch (err) {
+                    throw new Error(
+                        `Could not detect project slug from "name" field in package.json: ${err.message}`
+                    )
+                }
+            }
+
             const options = {
                 buildDirectory,
+                // Avoid setting message if it's blank, so that it doesn't override the default
                 ...(message ? {message} : undefined),
                 projectSlug,
                 target,