+++ packages/pwa-kit-react-sdk/src/ssr/server/react-rendering.js
@@ -228,7 +228,7 @@ const renderAppHtml = (req, res, error, appData) => {
 const renderApp = (args) => {
     const {req, res, appStateError, App, appState, location, routes, config} = args
     const deviceType = detectDeviceType(req)
-    const extractor = new ChunkExtractor({statsFile: BUNDLES_PATH})
+    const extractor = new ChunkExtractor({statsFile: BUNDLES_PATH, publicPath: getAssetUrl()})
     const routerContext = {}
     const appData = {App, appState, location, routes, routerContext, deviceType, extractor}
 
@@ -255,12 +255,10 @@ const renderApp = (args) => {
     let bundles = []
     /* istanbul ignore next */
     if (extractor) {
-        // Clone elements with the correct bundle path.
         bundles = extractor.getScriptElements().map((el) =>
             React.cloneElement(el, {
                 ...el.props,
-                ...scriptProps,
-                src: el.props.src && getAssetUrl(el.props.src.slice(1))
+                ...scriptProps
             })
         )
     }