+++ packages/pwa-kit-runtime/src/ssr/server/express.test.js
@@ -174,14 +174,9 @@ describe('_createApp validates environment variables', () => {
 describe('SSRServer operation', () => {
     const savedEnvironment = Object.assign({}, process.env)
     const sandbox = sinon.createSandbox()
-    let server
 
     afterEach(() => {
         sandbox.restore()
-        if (server) {
-            server.close()
-            server = null
-        }
         nock.cleanAll()
     })
 
@@ -849,7 +844,10 @@ describe('SSRServer persistent caching', () => {
         })
     )
 
-    const errorCases = [{url: '/?type=500', status: 500}, {url: '/?type=400', status: 400}]
+    const errorCases = [
+        {url: '/?type=500', status: 500},
+        {url: '/?type=400', status: 400}
+    ]
 
     errorCases.forEach(({url, status}) => {
         test(`should not cache responses with ${status} status codes`, () => {
@@ -984,6 +982,15 @@ describe('getRuntime', () => {
         }
     })
 
+    const matchExceptFunctionValues = (obj) => {
+        const entries = Object.entries(obj)
+        const matchers = entries.map(([key, value]) => {
+            const matcher = typeof value === 'function' ? expect.any(Function) : value
+            return [key, matcher]
+        })
+        return Object.fromEntries(matchers)
+    }
+
     const cases = [
         {
             env: {},
@@ -1017,11 +1024,11 @@ describe('getRuntime', () => {
         'should return a remote/development runtime $msg',
         ({env, expectedRuntime}) => {
             process.env = {...process.env, ...env}
-            expect(getRuntime()).toBe(expectedRuntime)
+            expect(getRuntime()).toMatchObject(matchExceptFunctionValues(expectedRuntime))
         }
     )
 
-    test('should return a remote/development runtime with the correct context', () => {
+    test('should return a remote/development runtime bound to the correct context', () => {
         const mockDevRuntime = getRuntime()
         const func = mockDevRuntime.returnMyName
         expect(func()).toBe(MockDevServerFactory.name)